Zabezpieczenie zarządzania aukcjami

+++ src/AppBundle/Controller/AuctionController.php
@@ -11,6 +11,7 @@ use Symfony\Component\Form\FormError;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 
 class AuctionController extends Controller
@@ -87,6 +88,12 @@ class AuctionController extends Controller
      */
     public function addAction(Request $request)
     {
+        /* KAŻDA Z PONIŻSZYCH TRZECH LINII KODU ODPOWIADA ZA SPRAWDZENIE, CZY UŻYTKOWNIK JEST ZALOGOWANY
+        if ($this->getUser() instanceof User)
+        if ($this->isGranted("ROLE_USER"))
+        */
+        $this->denyAccessUnlessGranted("ROLE_USER");
+
         $auction = new Auction();
 
         $form = $this->createForm(AuctionType::class, $auction);
@@ -100,7 +107,8 @@ class AuctionController extends Controller
 
             if ($form->isValid()) {
                 $auction
-                    ->setStatus(Auction::STATUS_ACTIVE);
+                    ->setStatus(Auction::STATUS_ACTIVE)
+                    ->setOwner($this->getUser());
 
                 $entityManager = $this->getDoctrine()->getManager();
                 $entityManager->persist($auction);
@@ -126,6 +134,12 @@ class AuctionController extends Controller
      */
     public function editAction(Request $request, Auction $auction)
     {
+        $this->denyAccessUnlessGranted("ROLE_USER");
+
+        if ($this->getUser() !== $auction->getOwner()) {
+            throw new AccessDeniedException();
+        }
+
         $form = $this->createForm(AuctionType::class, $auction);
 
         if ($request->isMethod("post")) {
@@ -152,6 +166,12 @@ class AuctionController extends Controller
      */
     public function deleteAction(Auction $auction)
     {
+        $this->denyAccessUnlessGranted("ROLE_USER");
+
+        if ($this->getUser() !== $auction->getOwner()) {
+            throw new AccessDeniedException();
+        }
+
         $entityManager = $this->getDoctrine()->getManager();
         $entityManager->remove($auction);
         $entityManager->flush();
@@ -170,6 +190,12 @@ class AuctionController extends Controller
      */
     public function finishAction(Auction $auction)
     {
+        $this->denyAccessUnlessGranted("ROLE_USER");
+
+        if ($this->getUser() !== $auction->getOwner()) {
+            throw new AccessDeniedException();
+        }
+
         $auction
             ->setExpiresAt(new \DateTime())
             ->setStatus(Auction::STATUS_FINISHED);
